Refactors Vector RBAC to use vector-config
Refactors the Vector agent and aggregator configurations to rely on the `vector-config` chart for managing RBAC resources and service accounts.

This change disables the creation of RBAC resources and service accounts within the `vector-agent` and `vector` charts, instead leveraging the existing resources managed by the `vector-config` chart. This ensures consistency and simplifies management of RBAC permissions across Vector deployments.

Updates the aggregator service account namespace to `observability` to align with the `vector-config` deployment.